hallucination_eval/requirements.txt
# Requirements for standalone hallucination evaluation scripts
google-generativeai>=0.3.0
tqdm>=4.64.0
python-dotenv>=0.19.0
requests>=2.28.0
